requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
transformers==4.35.0
trimesh==4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
xatlas==0.0.9
moderngl==5.10.0
torch==2.3.1
numpy==1.26.4
boto3==1.34.161